Start moving recv zbuf operations one abstraction above
