+++ include/zenoh-pico/system/common.h
@@ -69,8 +69,12 @@ _zn_socket_result_t _zn_create_udp_socket(const char *addr, int port, int recv_t
 int _zn_send_dgram_to(_zn_socket_t sock, const _z_wbuf_t *wbf, const struct sockaddr *dest, socklen_t salen);
 int _zn_recv_dgram_from(_zn_socket_t sock, _z_zbuf_t *zbf, struct sockaddr *from, socklen_t *salen);
 
+int _zn_send_bytes(_zn_socket_t sock, const uint8_t *buf, size_t len);
+int _zn_recv_bytes(_zn_socket_t sock, uint8_t *buf, size_t len);
+int _zn_recv_exact_bytes(_zn_socket_t sock, uint8_t *buf, size_t len);
+
 int _zn_send_wbuf(_zn_socket_t sock, const _z_wbuf_t *wbf);
 int _zn_recv_zbuf(_zn_socket_t sock, _z_zbuf_t *zbf);
-int _zn_recv_bytes(_zn_socket_t sock, uint8_t *buf, size_t len);
+int _zn_recv_exact_zbuf(_zn_socket_t sock, _z_zbuf_t *zbf, size_t len);
 
 #endif /* _ZENOH_PICO_SYSTEM_PRIVATE_COMMON_H */
